requirements.txt
numpy==1.26.1
pandas==2.1.1
python-dotenv==1.0.0
streamlit
